backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
pydantic==2.9.2
typing_extensions==4.12.2
pytest==8.2.2
httpx==0.27.0
